+++ modules/iphone/quHaoIos/Mine/CurrentDetailController.m
@@ -172,7 +172,7 @@
 
 - (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    int row = [indexPath row];
+    NSInteger row = [indexPath row];
     if(row==0||row==5||row==6){
         return 100;
     }else if(row==2||row==3||row==4){
@@ -334,7 +334,7 @@
 -(void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath
 {
     [tableView deselectRowAtIndexPath:indexPath animated:YES];
-    int row = [indexPath row];
+    NSInteger row = [indexPath row];
     if (row ==3) {
         [self CallPhone];
     }else if(row==2){
