+++ modules/iphone/quHaoIos/home/CityViewController.m
@@ -244,7 +244,7 @@
     int i,j;
     NSString *key;
     NSArray *value;
-    int count = [keys count];
+    NSInteger count = [keys count];
     for (i = 0; i < count; i++)
     {
         key = [keys objectAtIndex: i];
@@ -302,7 +302,7 @@
                  NSArray * value;
                  int i ;
                  int j ;
-                 int count = [keys count];
+                 NSInteger count = [keys count];
                  for (i = 0; i < count; i++)
                  {
                      value = [cities objectForKey: [keys objectAtIndex: i]];
@@ -330,7 +330,7 @@
                  NSArray * value;
                  int i ;
                  int j ;
-                 int count = [keys count];
+                 NSInteger count = [keys count];
                  for (i = 0; i < count; i++)
                  {
                      value = [cities objectForKey: [keys objectAtIndex: i]];
