+++ modules/iphone/quHaoIos/home/ListViewController.m
@@ -137,7 +137,7 @@
 -(void)tableView:(UITableView *)tableView didSelectRowAtIndexPath:(NSIndexPath *)indexPath
 {
     [tableView deselectRowAtIndexPath:indexPath animated:YES];
-    int row = [indexPath row];
+    NSInteger row = [indexPath row];
     MerchartModel *n = [_merchartsArray objectAtIndex:row];
     if (n)
     {
@@ -176,7 +176,7 @@
               [Helper showHUD2:@"服务器错误" andView:self.view andSize:100];
           }else{
               NSMutableArray *newMerc = [self addAfterInfo:jsonObjects];
-              int count = [newMerc count];
+              NSInteger count = [newMerc count];
               _allCount += count;
               if (count < 10)
               {
